T-G2: FPS selection in prep_pano.sh + FPS×error hybrid experiment
(a) prep_pano.sh gains an FPS mode: `prep_pano.sh <S> <insv> <N> fps [POOLX]`
    stitches a POOLX*N pool -> 1-crop coarse VGGT -> fps_select.py (farthest-point
    on camera poses) -> keeps N panos -> normal crop/curate/VGGT. Validated
    end-to-end on scene_023: 270 pool -> 90 FPS panos + 270 crops + full sparse/0.
    Also fixed two latent bugs surfaced by this path:
      - pano_to_perspective.py --directions parsed yaw/pitch as float -> %d
        filename crash; now int.
      - prep_pano.sh step-3 VGGT mounted weights at /wsrc but demo_colmap loads
        /wcache/model.pt -> mount weights at /wcache (matches step-1b).

(b) select_compare.py gains `hybrid` (FPS seed for coverage + remaining budget by
    error x novelty). 4-way (same 12 test / 78 budget, fused T-F8):
      uniform 18.94 | fps 19.55 | adaptive 18.16 | hybrid 19.17
    Honest result: pure FPS still wins; hybrid beats uniform/adaptive but not
    pure FPS; adaptive is unstable. Adding error-guidance doesn't help (high-error
    regions are intrinsically hard, not under-sampled). Recommendation: pure FPS.

fps_select.py added. tasks.md (T-G2) + OPTIMIZATION_PLAN.md updated.

+++ pano_pipeline/pano_to_perspective.py
@@ -302,7 +302,7 @@ def main():
     if args.directions:
         directions = []
         for pair in args.directions.split():
-            yaw, pitch = map(float, pair.split(","))
+            yaw, pitch = map(int, pair.split(","))  # int degrees (filename uses %d)
             directions.append((yaw, pitch))
     else:
         directions = PRESETS[args.preset]
